UnknownBlock; smithery materials & UI updates
Add UnknownBlock to act as a disguise for staged block-entities and defer loot/tool checks to the real block, fixing hidden-block drops/harvest behavior. Bump mod version and smithery API floor (gradle.properties + mods.toml) to 3.85.2 / 1.19.0. Restrict magic-only materials to the arcane focus via PartEligibility and retune several magic materials (harvest level, attack, durability, bow stats). Major refactor of MaterialsCatalogScreen: detail pane, icon hitboxes, parts caching, Index for reverse recipe lookups, richer detail lines (traits by scope, parts, sources, synergies), improved layout and performance. Small ModBlocks change to register UNKNOWN_BLOCK as UnknownBlock. These changes improve UX, correctness of hidden blocks, and Smithery integration.

+++ src/main/java/com/soul/soa_additions/oresight/client/OreSightIconRenderer.java
@@ -90,7 +90,7 @@ public final class OreSightIconRenderer implements IClientMobEffectExtensions {
         if (source.isEmpty()) return null;
         if (source.size() == 1) return source.get(0);
         long t = Minecraft.getInstance().level == null ? 0 : Minecraft.getInstance().level.getGameTime();
-        int idx = (int) Math.floorMod(t / CYCLE_TICKS, source.size());
+        int idx = Math.floorMod(t / CYCLE_TICKS, source.size());
         return source.get(idx);
     }
 }
